pyproject.toml
[build-system]
requires = ["flit_core >= 3.12.0, <4"]
build-backend = "flit_core.buildapi"

[project]
name = "gdm_robotics"
version = "1.0.1"
description = "Google DeepMind Robotics Interfaces and utilities."
requires-python = ">=3.10"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Typing :: Typed",
]
keywords = ["RL", "Reinforcement Learning", "AI", "ML"]
authors = [
  {name = "Google DeepMind", email = "robotics+oss@google.com"},
]
maintainers = [
  {name = "Claudio Fantacci", email = "cfantacci@google.com"},
  {name = "Francesco Romano", email = "fraromano@google.com"},
]

dependencies = [
  "numpy >= 2.0, < 3.0",
  "absl-py",
  "dm-env",
  "typing-extensions",
  "pytest",
  "gymnasium >= 1.0.0"
]

[project.urls]
Repository = "https://github.com/google-deepmind/gdm_robotics.git"
Issues = "https://github.com/google-deepmind/gdm_robotics/issues"


[tool.pytest.ini_options]
minversion = "6.2.5"
addopts = "-ra -q"
python_files = ["*_test.py"]
